requirements.txt
google-genai>=1.43.0
python-dotenv==1.1.0
sqlalchemy>=2.0.0
pydantic>=2.0.0
streamlit>=1.28.0
